Run test_emrun suite sequentially. In order to do that, it must be moved to its own file.
